Python Streaming Transactions script/requirements.txt
azure-eventhub==5.11.7
python-dotenv==1.0.1
